+++ backend/core/src/auth/controllers/user.controller.ts
@@ -77,11 +77,40 @@ export class UserController {
     )
   }
 
+  @Post("resend-partner-confirmation")
+  @UseGuards(OptionalAuthGuard, AuthzGuard)
+  @ApiOperation({
+    summary: "Resend partner confirmation",
+    operationId: "resendPartnerConfirmation",
+  })
+  async requestConfirmationResend(@Body() dto: EmailDto): Promise<StatusDto> {
+    await this.userService.resendPartnerConfirmation(dto)
+    return mapTo(StatusDto, { status: "ok" })
+  }
+
+  @Post("is-confirmation-token-valid")
+  @UseGuards(OptionalAuthGuard, AuthzGuard)
+  @ApiOperation({
+    summary: "Verifies token is valid",
+    operationId: "isUserConfirmationTokenValid",
+  })
+  async isUserConfirmationTokenValid(@Body() dto: ConfirmDto): Promise<boolean> {
+    return await this.userService.isUserConfirmationTokenValid(dto)
+  }
+
   @Post("resend-confirmation")
   @UseGuards(OptionalAuthGuard, AuthzGuard)
   @ApiOperation({ summary: "Resend confirmation", operationId: "resendConfirmation" })
   async confirmation(@Body() dto: EmailDto): Promise<StatusDto> {
     await this.userService.resendPublicConfirmation(dto)
+    return mapTo(StatusDto, { status: "ok" })
+  }
+
+  @Post("resend-partner-confirmation")
+  @UseGuards(OptionalAuthGuard, AuthzGuard)
+  @ApiOperation({ summary: "Resend confirmation", operationId: "resendPartnerConfirmation" })
+  async resendPartnerConfirmation(@Body() dto: EmailDto): Promise<StatusDto> {
+    await this.userService.resendPartnerConfirmation(dto)
     return mapTo(StatusDto, { status: "ok" })
   }